renamed go package/folder to later fix golint error
