+++ tests/gotests/doc.go
@@ -1,2 +1,2 @@
 // Package go contains automated tests written in Golang.
-package _go
+package gotests
